% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanosl.R
\name{almm}
\alias{almm}
\title{Fit an ALMM model to OSL data}
\usage{
almm(data, logged = F, sigma = NULL, messages = T, ...)
}
\arguments{
\item{data}{data[,1] are the dose values, data[,2] are the standard errors}

\item{logged}{specifies if the supplied doses (and associated standard errors) are logged (logged = T)}

\item{sigma}{specifies a fixed overdispersion valued (optional)}

\item{messages}{print status messages to the console (messages = T)}

\item{...}{additional arguments passed to \code{\link[rstan]{stan}}}
}
\value{
a fitted Stan object
}
\description{
This function fits an OSL Asymmetric Laplacian Mixture Model (ALMM) using Stan.
}
\seealso{
\code{\link{rstanosl}}
}
