% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanosl.R
\name{mxdm3}
\alias{mxdm3}
\title{Fit a MXDM3 model to OSL data}
\usage{
mxdm3(data, logged = F, sigma = NULL, messages = T, ...)
}
\arguments{
\item{data}{data[,1] are the dose values, data[,2] are the standard errors}

\item{logged}{specifies if the supplied doses (and associated standard errors) are logged (logged = T)}

\item{sigma}{specifies a fixed overdispersion value (optional - if set to NULL the parameter is estimated)}

\item{messages}{print status messages to the console (messages = T)}

\item{...}{additional arguments passed to \code{\link[rstan]{stan}}}
}
\value{
a fitted Stan object
}
\description{
This function fits an OSL 3 parameter Maximum Dose Model (MXDM3) using Stan.
}
\seealso{
\code{\link{rstanosl}}
}
