% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-osl_data.R
\docType{data}
\name{osl_data}
\alias{osl_data}
\title{OSL Example Data}
\format{
A data frame with the following variables:
\describe{
\item{Depth}{numeric: depth layer}
\item{De}{numeric: dose measurement}
\item{rel.err}{numeric: relative error for the dose}
\item{dose_rate}{numeric: constant dose rate per depth layer}
}
}
\source{
Synthetic OSL data
}
\usage{
data(osl_data)
}
\description{
This dataset provides an example of OSL data for the rstanosl package.
}
