% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-package.R
\docType{package}
\name{rstanosl}
\alias{rstanosl}
\alias{_PACKAGE}
\title{rstanosl: OSL model fitting with Stan}
\description{
Fit optical stimulated luminescence (OSL) data to statistical models with Stan.
}
\references{
Yates, L. A., Aandahl, Z., Brook, B. W., Jacobs, Z., Li, B., David, B. & Roberts, R. (2023). A new OSL dose model to account for post-depositional mixing of sediments (in prep)
}
\author{
Zach Aandahl \email{zach.aandahl@utas.edu.au}, Luke Yates \email{luke.yates@utas.edu.au}
}
\keyword{package}
